+++ grafana-plugin/integration-tests/utils/integrations.ts
@@ -1,5 +1,5 @@
 import { Page } from '@playwright/test';
-import { clickButton } from './forms';
+import { clickButton, selectDropdownValue } from './forms';
 import { goToOnCallPage } from './navigation';
 
 const CREATE_INTEGRATION_MODAL_TEST_ID_SELECTOR = 'div[data-testid="create-integration-modal"]';
@@ -31,27 +31,47 @@ export const createIntegration = async (page: Page, integrationName: string): Pr
   await grafanaUpdateBtn.click();
 };
 
+export const assignEscalationChainToIntegration = async (page: Page, escalationChainName: string): Promise<void> => {
+  await page.getByTestId('integration-escalation-chain-not-selected').click();
+
+  // assign the escalation chain to the integration
+  await selectDropdownValue({
+    page,
+    selectType: 'grafanaSelect',
+    placeholderText: 'Select Escalation Chain',
+    value: escalationChainName,
+    startingLocator: page.getByTestId('integration-block-item'),
+  });
+};
+
+export const sendDemoAlert = async (page: Page): Promise<void> => {
+  await clickButton({ page, buttonText: 'Send demo alert', dataTestId: 'send-demo-alert' });
+  await clickButton({ page, buttonText: 'Send Alert', dataTestId: 'submit-send-alert' });
+};
+
 export const createIntegrationAndSendDemoAlert = async (
   page: Page,
   integrationName: string,
-  _escalationChainName: string
+  escalationChainName: string
 ): Promise<void> => {
   await createIntegration(page, integrationName);
+  await assignEscalationChainToIntegration(page, escalationChainName);
+  await sendDemoAlert(page);
+};
 
-  /*
-   * TODO: This is slightly more complicated now, change this in next iteration */
-  // const integrationSettingsElement = page.getByTestId('integration-settings');
+export const filterIntegrationsTableAndGoToDetailPage = async (page: Page, integrationName: string): Promise<void> => {
+  // filter the integrations page by the integration in question, then go to its detail page
+  await selectDropdownValue({
+    page,
+    selectType: 'grafanaSelect',
+    placeholderText: 'Search or filter results...',
+    value: integrationName,
+    pressEnterInsteadOfSelectingOption: true,
+  });
 
-  // // assign the escalation chain to the integration
-  // await selectDropdownValue({
-  //   page,
-  //   selectType: 'grafanaSelect',
-  //   placeholderText: 'Select Escalation Chain',
-  //   value: escalationChainName,
-  //   startingLocator: integrationSettingsElement,
-  // });
-
-  // send demo alert
-  await clickButton({ page, buttonText: 'Send demo alert', dataTestId: 'send-demo-alert' });
-  await clickButton({ page, buttonText: 'Send Alert', dataTestId: 'submit-send-alert' });
+  await (
+    await page.waitForSelector(
+      `div[data-testid="integrations-table"] table > tbody > tr > td:first-child a >> text=${integrationName}`
+    )
+  ).click();
 };